requirements.txt
fastapi
uvicorn

openai
tiktoken

python-dotenv
python-multipart